requirements.txt
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==3.0.1
dnspython==2.3.0
eventlet==0.33.3
greenlet==2.0.2
idna==3.4
PyQt6==6.4.1
PyQt6-Qt6==6.4.2
PyQt6-sip==13.4.1
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.2
six==1.16.0
urllib3==1.26.14
websocket-client==1.5.0

transliterate==1.10.2